Refactor: Split oversized files into modular, single-responsibility components
- Split bt_task/logic.rs into logic.rs, choker.rs, dht.rs, lpd.rs, and tracker.rs to comply with 400-line hard cap.
- Split orchestrator/structs.rs into state.rs, command.rs, worker_command.rs, event.rs, subtask_event.rs, and vpn_enforcement.rs.
- Resolved module exports and cascading import updates.
- Verified all unit and integration tests via make green-loop.
