+++ aura-core/src/bt_task/logic.rs
@@ -4,11 +4,10 @@ use crate::bitfield::Bitfield;
 use crate::peer_registry::PeerRegistry;
 use crate::piece_picker::PiecePicker;
 use crate::torrent::Torrent;
-use crate::tracker::{Peer, TrackerClient};
+use crate::tracker::Peer;
 use crate::{Error, InfoHash, Result, TaskId};
 use std::sync::Arc;
 use tokio::sync::{mpsc, Mutex};
-use tracing::{debug, info, warn};
 
 #[derive(Debug)]
 pub struct BtTaskState {
@@ -116,125 +115,6 @@ impl BtTask {
         }
     }
 
-    pub async fn run_dht_loop(&self, token: tokio_util::sync::CancellationToken) -> Result<()> {
-        let info_hash = self.state.info_hash;
-        loop {
-            tokio::select! {
-                _ = token.cancelled() => break,
-                _ = async {
-                    let (tx, mut rx) = tokio::sync::mpsc::channel(1);
-                    let cmd = crate::dht::DhtCommand::GetPeers {
-                        info_hash,
-                        reply_tx: tx,
-                    };
-
-                    if let Err(e) = self.dht_tx.send(cmd).await {
-                        warn!("Failed to send DHT command: {}", e);
-                        return;
-                    }
-
-                    if let Some(addrs) = rx.recv().await {
-                        let mut peers = Vec::new();
-                        for addr in addrs {
-                            let ip: std::net::IpAddr = addr.ip();
-                            peers.push(Peer {
-                                id: None,
-                                ip: ip.to_string(),
-                                port: addr.port(),
-                            });
-                        }
-
-                        if !peers.is_empty() {
-                            info!(%self.id, count = peers.len(), "Discovered peers from DHT");
-                            let mut registry = self.state.registry.lock().await;
-                            let added = registry.add_peers(peers);
-                            debug!(%self.id, added, "Added unique DHT peers to registry");
-                        }
-                    }
-                } => {}
-            }
-
-            tokio::select! {
-                _ = token.cancelled() => break,
-                _ = tokio::time::sleep(std::time::Duration::from_secs(120)) => {}
-            }
-        }
-        Ok(())
-    }
-
-    pub async fn run_lpd_loop(
-        &self,
-        port: u16,
-        token: tokio_util::sync::CancellationToken,
-    ) -> Result<()> {
-        let info_hash = self.state.info_hash;
-        info!(%self.id, "Starting LPD announcement loop");
-
-        // Initial announcement
-        let _ = self
-            .lpd_tx
-            .send(crate::lpd::LpdCommand::Announce { info_hash, port })
-            .await;
-
-        loop {
-            tokio::select! {
-                _ = token.cancelled() => {
-                    let _ = self
-                        .lpd_tx
-                        .send(crate::lpd::LpdCommand::Remove {
-                            info_hash,
-                        })
-                        .await;
-                    break;
-                }
-                _ = tokio::time::sleep(std::time::Duration::from_secs(300)) => {
-                }
-            }
-        }
-        Ok(())
-    }
-
-    pub async fn run_tracker_loop(
-        &self,
-        my_id: [u8; 20],
-        port: u16,
-        token: tokio_util::sync::CancellationToken,
-        local_addr: Option<std::net::IpAddr>,
-        user_agent: Option<String>,
-        proxy: Option<String>,
-    ) -> Result<()> {
-        let tracker = TrackerClient::new(my_id, port, local_addr, user_agent, proxy);
-        info!(%self.id, "Starting tracker announce");
-
-        loop {
-            let torrent_opt = self.state.torrent.lock().await.clone();
-            if let Some(torrent) = torrent_opt {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    res = tracker.announce(&torrent) => {
-                        match res {
-                            Ok(peers) => {
-                                info!(%self.id, count = peers.len(), "Discovered peers from tracker");
-                                let mut registry = self.state.registry.lock().await;
-                                let added = registry.add_peers(peers);
-                                debug!(%self.id, added, "Added unique peers to registry");
-                            }
-                            Err(e) => {
-                                tracing::warn!(%self.id, error = %e, "All tracker announces failed");
-                            }
-                        }
-                    }
-                }
-            }
-
-            tokio::select! {
-                _ = token.cancelled() => break,
-                _ = tokio::time::sleep(std::time::Duration::from_secs(60)) => {}
-            }
-        }
-        Ok(())
-    }
-
     /// Picks a peer to connect to and optionally a piece to download.
     pub async fn pick_work(&self) -> Option<(Option<usize>, Peer)> {
         let mut registry = self.state.registry.lock().await;
@@ -272,113 +152,6 @@ impl BtTask {
         let mut registry = self.state.registry.lock().await;
         if let Some(ps) = registry.get_mut(addr) {
             ps.peer_interested = interested;
-        }
-    }
-
-    pub async fn run_choker_loop(
-        &self,
-        worker_cmd_tx: tokio::sync::broadcast::Sender<crate::orchestrator::WorkerCommand>,
-        token: tokio_util::sync::CancellationToken,
-    ) -> Result<()> {
-        let mut ticks = 0;
-        let mut interval = tokio::time::interval(std::time::Duration::from_secs(10));
-        loop {
-            tokio::select! {
-                _ = token.cancelled() => break,
-                _ = interval.tick() => {}
-            }
-
-            ticks += 1;
-            let optimistic = ticks % 3 == 0;
-            self.tick_choker(optimistic, &worker_cmd_tx).await;
-        }
-        Ok(())
-    }
-
-    pub async fn tick_choker(
-        &self,
-        optimistic: bool,
-        worker_cmd_tx: &tokio::sync::broadcast::Sender<crate::orchestrator::WorkerCommand>,
-    ) {
-        let mut registry = self.state.registry.lock().await;
-        registry.tick_rates(10.0);
-
-        if optimistic {
-            registry.reset_optimistic_unchokes();
-        }
-
-        let is_seeding = self
-            .state
-            .bitfield
-            .lock()
-            .await
-            .as_ref()
-            .map(|bf| bf.is_complete())
-            .unwrap_or(false);
-
-        let mut peers: Vec<_> = registry.get_all_connected().into_iter().collect();
-
-        // Shuffle peers first so that stable sort keeps tied peers in random order
-        use rand::seq::SliceRandom;
-        peers.shuffle(&mut rand::rng());
-
-        // Sort peers:
-        // If we are leeching: unchoke peers we download fastest from.
-        // If we are seeding: unchoke peers we upload fastest to.
-        peers.sort_by(|a, b| {
-            if is_seeding {
-                b.upload_rate.partial_cmp(&a.upload_rate)
-            } else {
-                b.download_rate.partial_cmp(&a.download_rate)
-            }
-            .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        let mut unchoked_count = 0;
-        let mut candidates_for_optimistic = Vec::new();
-
-        for p in peers.iter_mut() {
-            // tit-for-tat unchokes the top 4 INTERESTED peers.
-            if unchoked_count < 4 && p.peer_interested {
-                if p.am_choking {
-                    p.am_choking = false;
-                    let _ = worker_cmd_tx.send(crate::orchestrator::WorkerCommand::Unchoke(
-                        p.peer.ip.clone(),
-                        p.peer.port,
-                    ));
-                }
-                unchoked_count += 1;
-            } else {
-                if p.is_optimistic_unchoke {
-                    unchoked_count += 1;
-                    continue;
-                }
-
-                if p.am_choking {
-                    if p.peer_interested {
-                        candidates_for_optimistic.push(p);
-                    }
-                } else {
-                    // This peer was unchoked but is no longer in the top 4.
-                    p.am_choking = true;
-                    let _ = worker_cmd_tx.send(crate::orchestrator::WorkerCommand::Choke(
-                        p.peer.ip.clone(),
-                        p.peer.port,
-                    ));
-                }
-            }
-        }
-
-        // Handle Optimistic Unchoke
-        if optimistic && !candidates_for_optimistic.is_empty() {
-            // Since candidates are already shuffled, we can just pick the first one
-            let p = &mut candidates_for_optimistic[0];
-            p.am_choking = false;
-            p.is_optimistic_unchoke = true;
-            let _ = worker_cmd_tx.send(crate::orchestrator::WorkerCommand::Unchoke(
-                p.peer.ip.clone(),
-                p.peer.port,
-            ));
         }
     }
