+++ aura-core/src/bt_task/mod.rs
@@ -1,4 +1,8 @@
+pub mod choker;
+pub mod dht;
 pub mod logic;
+pub mod lpd;
+pub mod tracker;
 pub use logic::*;
 
 #[cfg(test)]
