+++ aura-core/src/orchestrator/state.rs
@@ -1,0 +1,126 @@
+use super::{Command, Event, SubTaskEvent, WorkerCommand};
+use crate::bt_task::BtTask;
+use crate::buffer_pool::BufferPool;
+use crate::dht::DhtCommand;
+use crate::nat::NatCommand;
+use crate::task::MetaTask;
+use crate::throttler::Throttler;
+use crate::{InfoHash, TaskId};
+use arc_swap::ArcSwap;
+use rand::RngExt;
+use std::collections::HashMap;
+use std::sync::Arc;
+use tokio::sync::{broadcast, mpsc};
+use tokio_util::sync::CancellationToken;
+
+pub struct Orchestrator {
+    pub(crate) tasks: HashMap<TaskId, MetaTask>,
+    pub(crate) bt_registry: HashMap<InfoHash, TaskId>,
+    pub(crate) bt_tasks: HashMap<TaskId, Arc<BtTask>>,
+    pub(crate) worker_command_txs: HashMap<TaskId, tokio::sync::broadcast::Sender<WorkerCommand>>,
+    pub(crate) cancellation_tokens: HashMap<TaskId, CancellationToken>,
+    pub(crate) command_rx: mpsc::Receiver<Command>,
+    pub(crate) event_tx: broadcast::Sender<Event>,
+    pub(crate) storage_tx: mpsc::Sender<crate::storage::StorageRequest>,
+    pub(crate) storage_completion_rx: mpsc::Receiver<crate::storage::StorageEvent>,
+    pub(crate) subtask_tx: mpsc::Sender<SubTaskEvent>,
+    pub(crate) subtask_rx: mpsc::Receiver<SubTaskEvent>,
+    pub(crate) dht_tx: mpsc::Sender<DhtCommand>,
+    pub(crate) lpd_tx: mpsc::Sender<crate::lpd::LpdCommand>,
+    pub(crate) scrub_tx: mpsc::Sender<crate::scrubber::ScrubberCommand>,
+    pub(crate) scrub_rx: Option<mpsc::Receiver<crate::scrubber::ScrubberCommand>>,
+    pub(crate) _nat_tx: mpsc::Sender<NatCommand>,
+    pub(crate) peer_id: [u8; 20],
+    pub(crate) throttler: Arc<Throttler>,
+    pub(crate) vpn_provider: Option<Arc<dyn crate::vpn::VpnProvider>>,
+    pub(crate) vpn_watch_tx: tokio::sync::watch::Sender<Option<Arc<dyn crate::vpn::VpnProvider>>>,
+    pub(crate) config: Arc<ArcSwap<crate::Config>>,
+    pub(crate) power_manager: crate::power::PowerManager,
+    pub(crate) hook_manager: crate::hooks::HookManager,
+    pub(crate) credential_provider: Arc<crate::config::credentials::CredentialProvider>,
+    pub(crate) dns_resolver: Arc<crate::net_util::TokioResolver>,
+    pub(crate) pool: BufferPool,
+    pub(crate) db: sled::Db,
+    pub(crate) hsts_cache: crate::security::HstsCache,
+}
+
+impl Orchestrator {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        command_rx: mpsc::Receiver<Command>,
+        storage_tx: mpsc::Sender<crate::storage::StorageRequest>,
+        storage_completion_rx: mpsc::Receiver<crate::storage::StorageEvent>,
+        dht_tx: mpsc::Sender<DhtCommand>,
+        lpd_tx: mpsc::Sender<crate::lpd::LpdCommand>,
+        nat_tx: mpsc::Sender<NatCommand>,
+        config: Arc<ArcSwap<crate::Config>>,
+        pool: BufferPool,
+        db: sled::Db,
+        dns_resolver: Arc<crate::net_util::TokioResolver>,
+    ) -> (Self, broadcast::Sender<Event>) {
+        let (event_tx, _event_rx) = broadcast::channel(1024);
+        let (subtask_tx, subtask_rx) = mpsc::channel(4096);
+        let (scrub_tx, scrub_rx) = mpsc::channel(1024);
+
+        let mut peer_id = [0u8; 20];
+        peer_id[..8].copy_from_slice(b"-AR0001-");
+        rand::rng().fill(&mut peer_id[8..]);
+
+        let initial_config = config.load();
+        let throttler = Arc::new(Throttler::new(
+            initial_config.bandwidth.global_download_limit,
+            initial_config.bandwidth.global_upload_limit,
+        ));
+
+        let vpn_provider = Self::create_vpn_provider(&initial_config);
+        let (vpn_watch_tx, _vpn_watch_rx) = tokio::sync::watch::channel(vpn_provider.clone());
+        let hook_manager = crate::hooks::HookManager::new(initial_config.hooks.clone());
+
+        let mut credential_provider = crate::config::credentials::CredentialProvider::new();
+        if let Some(ref netrc) = initial_config.credentials.netrc_path {
+            if let Err(e) = credential_provider.load_netrc(netrc) {
+                tracing::warn!("Failed to load .netrc from {}: {}", netrc, e);
+            }
+        }
+        if let Some(ref cookie_file) = initial_config.credentials.cookie_file {
+            if let Err(e) = credential_provider.load_cookies(cookie_file) {
+                tracing::warn!("Failed to load cookies from {}: {}", cookie_file, e);
+            }
+        }
+        let credential_provider = Arc::new(credential_provider);
+
+        (
+            Self {
+                tasks: HashMap::new(),
+                bt_registry: HashMap::new(),
+                bt_tasks: HashMap::new(),
+                worker_command_txs: HashMap::new(),
+                cancellation_tokens: HashMap::new(),
+                command_rx,
+                event_tx: event_tx.clone(),
+                storage_tx,
+                storage_completion_rx,
+                subtask_tx,
+                subtask_rx,
+                scrub_tx,
+                scrub_rx: Some(scrub_rx),
+                dht_tx,
+                lpd_tx,
+                _nat_tx: nat_tx,
+                peer_id,
+                throttler,
+                vpn_provider,
+                vpn_watch_tx,
+                config,
+                power_manager: crate::power::PowerManager::new(),
+                hook_manager,
+                credential_provider,
+                dns_resolver,
+                pool,
+                db,
+                hsts_cache: crate::security::HstsCache::new(),
+            },
+            event_tx,
+        )
+    }
+}
