+++ aura-core/src/orchestrator/worker_command.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone)]
+pub enum WorkerCommand {
+    CancelPiece(usize),
+    RequestPiece(usize),
+    Choke(String, u16),
+    Unchoke(String, u16),
+}
